完善同步流程 Signed-off-by: xxonehjh <code234@qq.com>

+++ src/main/java/com/hjh/files/sync/client/ClientFolder.java
@@ -11,6 +11,8 @@ import com.hjh.files.sync.common.ILog;
 import com.hjh.files.sync.common.RemoteFile;
 import com.hjh.files.sync.common.RemoteFileFactory;
 import com.hjh.files.sync.common.RemoteFileManage;
+import com.hjh.files.sync.common.RemoteSyncConfig;
+import com.hjh.files.sync.common.StopAble;
 import com.hjh.files.sync.common.util.MD5;
 
 public class ClientFolder {
@@ -31,10 +33,8 @@ public class ClientFolder {
 		this.url = url;
 		this.cache = new File(store_folder, CLIENT_CACHE_FOLDER_NAME);
 		if (!this.cache.isDirectory()) {
-			Asserts.check(
-					this.cache.mkdir(),
-					"can not create cache folder for client on :"
-							+ this.cache.getAbsolutePath());
+			Asserts.check(this.cache.mkdir(),
+					"can not create cache folder for client on :" + this.cache.getAbsolutePath());
 		}
 	}
 
@@ -62,10 +62,9 @@ public class ClientFolder {
 		this.store_folder = store_folder;
 	}
 
-	public void sync() throws IOException {
+	public void sync(StopAble stop) throws IOException {
 
-		String store_path = new File(new File(store_folder), name)
-				.getCanonicalPath();
+		String store_path = new File(new File(store_folder), name).getCanonicalPath();
 		if (null == fromManage) {
 			fromManage = RemoteFileFactory.queryManage(url);
 		}
@@ -77,33 +76,39 @@ public class ClientFolder {
 		Asserts.check(root.isDirectory(), "must be a directory :" + store_path);
 		long time = System.currentTimeMillis();
 		try {
-			doSync(null, root);
+			if (stop.isStop()) {
+				return;
+			}
+			doSync(stop, null, root);
 		} finally {
 			long end = System.currentTimeMillis();
-			logger.stdout(String.format("sync finish[%s](cost: %s) %s => %s",
-					name, (end - time) / 1000 + "s", url, store_path));
+			logger.stdout(String.format("sync finish[%s](cost: %s) %s => %s", name, (end - time) / 1000 + "s", url,
+					store_path));
 		}
 	}
 
-	private void doSync(RemoteFile from, File target) throws IOException {
+	private void doSync(StopAble stop, RemoteFile from, File target) throws IOException {
 
 		if (null == from || from.isFolder()) { // 目录同步
 			String path = null == from ? null : from.path();
+			if (stop.isStop()) {
+				return;
+			}
 			RemoteFile[] remotes = fromManage.list(path);
+			if (stop.isStop()) {
+				return;
+			}
 			if (target.isFile()) {
-				logger.info("remove file:" + target.getAbsolutePath());
-				Asserts.check(target.delete(),
-						"delete file fail : " + target.getAbsolutePath());
+				logger.stdout("remove file:" + target.getAbsolutePath());
+				Asserts.check(target.delete(), "delete file fail : " + target.getAbsolutePath());
 			}
 			if (!target.exists()) {
-				logger.info(String.format("sync folder[%s] %s => %s", name,
-						path, target.getAbsolutePath()));
-				Asserts.check(target.mkdir(),
-						"create folder fail : " + target.getAbsolutePath());
+				logger.stdout(String.format("sync folder[%s] %s => %s", name, path, target.getAbsolutePath()));
+				Asserts.check(target.mkdir(), "create folder fail : " + target.getAbsolutePath());
 			}
 			String[] exists = target.list();
 			for (RemoteFile item : remotes) {
-				doSync(item, new File(target, item.name()));
+				doSync(stop, item, new File(target, item.name()));
 				if (null != exists) {
 					for (int i = 0; i < exists.length; i++) {
 						if (exists[i] != null && exists[i].equals(item.name())) {
@@ -118,28 +123,27 @@ public class ClientFolder {
 					if (exists[i] != null) {
 						File cur_exist = new File(target, exists[i]);
 						if (cur_exist.isDirectory()) {
-							logger.info("remove directory:"
-									+ cur_exist.getAbsolutePath());
+							logger.stdout("remove directory:" + cur_exist.getAbsolutePath());
 							FileUtils.deleteDirectory(cur_exist);
 						} else {
-							logger.info("remove file:"
-									+ cur_exist.getAbsolutePath());
-							Asserts.check(
-									cur_exist.delete(),
-									"can not delete file :"
-											+ cur_exist.getAbsolutePath());
+							logger.stdout("remove file:" + cur_exist.getAbsolutePath());
+							Asserts.check(cur_exist.delete(), "can not delete file :" + cur_exist.getAbsolutePath());
 						}
 					}
 				}
 			}
 		} else { // 文件同步
 			if (!isSame(from, target)) {
-				logger.info(String.format("sync file[%s] %s => %s", name,
-						from.path(), target.getAbsolutePath()));
+				logger.stdout(String.format("sync file[%s] %s => %s", name, from.path(), target.getAbsolutePath()));
+				if (stop.isStop()) {
+					return;
+				}
 				String md5 = fromManage.md5(from.path());
-
+				if (stop.isStop()) {
+					return;
+				}
 				if (target.isDirectory()) {
-					logger.info("remove directory:" + target.getAbsolutePath());
+					logger.stdout("remove directory:" + target.getAbsolutePath());
 					FileUtils.deleteDirectory(target);
 				}
 
@@ -149,22 +153,32 @@ public class ClientFolder {
 					if (!current_cache_root.exists()) {
 						current_cache_root.mkdir();
 					}
+					if (stop.isStop()) {
+						return;
+					}
 					int totalParts = fromManage.partCount(from.length());
+					if (stop.isStop()) {
+						return;
+					}
 					for (int i = 0; i < totalParts; i++) {
 						File cur_part = new File(current_cache_root, i + "");
 						if (!cur_part.exists()) {
+							if (stop.isStop()) {
+								return;
+							}
 							byte[] part_data = fromManage.part(from.path(), i);
-							logger.info(String.format(
-									"[%s] [%s] receive part data %d", name,
-									from.path(), part_data.length));
+							logger.debug(String.format("[%s] [%s] receive part data %d", name, from.path(),
+									part_data.length));
 							FileUtils.writeByteArrayToFile(cur_part, part_data);
+							if (stop.isStop()) {
+								return;
+							}
 						}
 					}
 
 					File target_temp = new File(current_cache_root, "temp");
 					if (target_temp.isDirectory()) {
-						logger.info("remove directory:"
-								+ target_temp.getAbsolutePath());
+						logger.stdout("remove directory:" + target_temp.getAbsolutePath());
 						FileUtils.deleteDirectory(target_temp);
 					}
 
@@ -172,8 +186,7 @@ public class ClientFolder {
 						String cache_md5 = MD5.md5(target_temp);
 						if (!md5.equals(cache_md5)) {
 							Asserts.check(target_temp.delete(),
-									"can not delete wrong file[md5 do not match]:"
-											+ target_temp.getAbsolutePath());
+									"can not delete wrong file[md5 do not match]:" + target_temp.getAbsolutePath());
 						}
 					}
 
@@ -181,36 +194,27 @@ public class ClientFolder {
 						target_temp.createNewFile();
 						for (int i = 0; i < totalParts; i++) {
 							File cur_part = new File(current_cache_root, i + "");
-							FileUtils.writeByteArrayToFile(target_temp,
-									FileUtils.readFileToByteArray(cur_part),
-									true);
+							FileUtils.writeByteArrayToFile(target_temp, FileUtils.readFileToByteArray(cur_part), true);
 						}
 					}
 
 					{
 						String cache_md5 = MD5.md5(target_temp);
 						if (!md5.equals(cache_md5)) {
-							logger.info("clear dirty directory : "
-									+ current_cache_root.getAbsolutePath());
+							logger.stdout("clear dirty directory : " + current_cache_root.getAbsolutePath());
 							FileUtils.deleteDirectory(current_cache_root);
-							throw new RuntimeException(
-									"can not fetch correct data from remote for:"
-											+ from.path());
+							throw new RuntimeException("can not fetch correct data from remote for:" + from.path());
 						}
 					}
 
 					if (target.isFile()) {
-						logger.info("remove unmatch file:"
-								+ target.getAbsolutePath());
-						Asserts.check(target.delete(), String.format(
-								"can not delete file : %s",
-								target.getAbsolutePath()));
+						logger.stdout("remove unmatch file:" + target.getAbsolutePath());
+						Asserts.check(target.delete(),
+								String.format("can not delete file : %s", target.getAbsolutePath()));
 					}
 
-					Asserts.check(target_temp.renameTo(target), String.format(
-							"can not move file: %s => %s",
-							target_temp.getAbsolutePath(),
-							target.getAbsolutePath()));
+					Asserts.check(target_temp.renameTo(target), String.format("can not move file: %s => %s",
+							target_temp.getAbsolutePath(), target.getAbsolutePath()));
 
 				}
 				target.setLastModified(from.lastModify());
@@ -234,9 +238,8 @@ public class ClientFolder {
 		}
 
 		if (from.lastModify() != root.lastModified()) {
-			logger.info(String.format("[%s] %d <> %d", from.path(),
-					from.lastModify(), root.lastModified()));
-			if (Math.abs(from.lastModify() - root.lastModified()) > 800) {
+			logger.debug(String.format("[%s] %d <> %d", from.path(), from.lastModify(), root.lastModified()));
+			if (Math.abs(from.lastModify() - root.lastModified()) > RemoteSyncConfig.getMinDiffTime()) {
 				return false;
 			}
 		}
